requirements.txt
Flask==3.1.2
flask_cors==6.0.1
cerebras_cloud_sdk==1.35.0
pydantic==2.11.7
python-dotenv==1.1.1
mem0ai
chromadb
sentence_transformers
ngrok